[PAY-3362] Move link resolver to common hook
