+++ packages/common/src/hooks/chats/index.ts
@@ -2,3 +2,4 @@ export * from './types'
 export * from './useCanSendMessage'
 export * from './useSetInboxPermissions'
 export * from './useTrackPlayer'
+export * from './useAudiusLinkResolver'
