+++ packages/web/src/pages/chat-page/components/ChatComposer.tsx
@@ -8,18 +8,10 @@ import {
   useEffect
 } from 'react'
 
-import { ID } from '@audius/common/models'
+import { useAudiusLinkResolver } from '@audius/common/hooks'
 import { chatActions } from '@audius/common/store'
-import {
-  decodeHashId,
-  formatCollectionName,
-  formatTrackName,
-  formatUserName,
-  matchAudiusLinks,
-  splitOnNewline
-} from '@audius/common/utils'
+import { splitOnNewline } from '@audius/common/utils'
 import { IconSend, IconButton, Text, TextProps } from '@audius/harmony'
-import { Track, Playlist, User, ResolveApi } from '@audius/sdk'
 import cn from 'classnames'
 import { useDispatch } from 'react-redux'
 
@@ -31,12 +23,6 @@ import styles from './ChatComposer.module.css'
 import { ComposerCollectionInfo, ComposerTrackInfo } from './ComposePreviewInfo'
 
 const { sendMessage } = chatActions
-const {
-  instanceOfTrackResponse,
-  instanceOfUserResponse,
-  instanceOfPlaylistResponse
-} = ResolveApi
-
 const messages = {
   sendMessage: 'Send Message',
   sendMessagePlaceholder: 'Start typing...'
@@ -88,67 +74,22 @@ export const ChatComposer = (props: ChatComposerProps) => {
   const [value, setValue] = useState(presetMessage ?? '')
   const [focused, setFocused] = useState(false)
 
-  // Maintain bidirectional maps of audius links to human readable format
-  const linkToHuman = useRef<{ [key: string]: string }>({}).current
-  const humanToData = useRef<{
-    [key: string]: { link: string; data: Track | Playlist | User }
-  }>({}).current
-
-  // The track id used to render the composer preview
-  const [trackId, setTrackId] = useState<ID | null>(null)
-  // The collection id used to render the composer preview
-  const [collectionId, setCollectionId] = useState<ID | null>(null)
-
   const ref = useRef<HTMLTextAreaElement>(null)
   const chatIdRef = useRef(chatId)
 
-  const resolveLinks = useCallback(
-    async (value: string) => {
-      setValue(value)
-
-      const { matches } = matchAudiusLinks({
-        text: value,
-        hostname: env.PUBLIC_HOSTNAME
-      })
-
-      const sdk = await audiusSdk()
-      for (const match of matches) {
-        if (!(match in linkToHuman)) {
-          const res = await sdk.resolve({ url: match })
-          if (res.data) {
-            if (instanceOfTrackResponse(res)) {
-              const human = formatTrackName({ track: res.data })
-              linkToHuman[match] = human
-              humanToData[human] = { link: match, data: res.data }
-            } else if (instanceOfPlaylistResponse(res)) {
-              const human = formatCollectionName({ collection: res.data[0] })
-              linkToHuman[match] = human
-              humanToData[human] = { link: match, data: res.data[0] }
-            } else if (instanceOfUserResponse(res)) {
-              const human = formatUserName({ user: res.data })
-              linkToHuman[match] = human
-              humanToData[human] = { link: match, data: res.data }
-            }
-          }
-        } else {
-          // If we already loaded the track, delay showing by 500ms
-          // to give the user some sense of what is happening.
-          await new Promise((resolve) => setTimeout(resolve, 500))
-        }
-      }
-
-      // Sort here to make sure that we replace all content before
-      // replacing their substrings.
-      let editedValue = value
-      for (const [link, human] of Object.entries(linkToHuman).sort(
-        (a, b) => b[0].length - a[0].length
-      )) {
-        editedValue = editedValue.replaceAll(link, human)
-      }
-      setValue(editedValue)
-    },
-    [setValue, linkToHuman, humanToData]
-  )
+  const {
+    trackId,
+    collectionId,
+    resolveLinks,
+    restoreLinks,
+    getMatches,
+    handleBackspace,
+    reset
+  } = useAudiusLinkResolver({
+    value,
+    hostname: env.PUBLIC_HOSTNAME,
+    audiusSdk
+  })
 
   useEffect(() => {
     if (presetMessage) {
@@ -158,45 +99,24 @@ export const ChatComposer = (props: ChatComposerProps) => {
 
   const handleChange = useCallback(
     async (e: ChangeEvent<HTMLTextAreaElement>) => {
-      resolveLinks(e.target.value)
+      setValue(e.target.value)
+      const editedValue = await resolveLinks(e.target.value)
+      setValue(editedValue)
     },
-    [resolveLinks]
+    [resolveLinks, setValue]
   )
-
-  useEffect(() => {
-    for (const [human, { data }] of Object.entries(humanToData)) {
-      if (value.includes(human)) {
-        if ('title' in data) {
-          setTrackId(decodeHashId(data.id))
-        } else if (
-          Array.isArray(data) &&
-          data.length > 0 &&
-          'playlistName' in data[0]
-        ) {
-          setCollectionId(decodeHashId(data.id))
-        }
-        return
-      }
-    }
-    setTrackId(null)
-    setCollectionId(null)
-  }, [trackId, humanToData, value])
 
   const handleSubmit = useCallback(
     async (e?: FormEvent) => {
       e?.preventDefault()
       if (chatId && value) {
         // On submit, actually send audius links rather than the human readable format
-        let editedValue = value
-        for (const [human, { link }] of Object.entries(humanToData)) {
-          editedValue = editedValue.replaceAll(human, link)
-        }
-        dispatch(sendMessage({ chatId, message: editedValue }))
+        dispatch(sendMessage({ chatId, message: restoreLinks(value) }))
         setValue('')
         onMessageSent()
       }
     },
-    [chatId, value, onMessageSent, humanToData, dispatch]
+    [chatId, value, onMessageSent, restoreLinks, dispatch]
   )
 
   // Submit when pressing enter while not holding shift
@@ -212,20 +132,17 @@ export const ChatComposer = (props: ChatComposerProps) => {
         const textarea = e.target as HTMLTextAreaElement
         const cursorPosition = textarea.selectionStart
         const textBeforeCursor = textarea.value.slice(0, cursorPosition)
-        const matched = Object.keys(humanToData).find((i) =>
-          textBeforeCursor.endsWith(i)
-        )
-        if (matched) {
+        const editValue = handleBackspace({
+          cursorPosition,
+          textBeforeCursor
+        })
+        if (editValue) {
           e.preventDefault()
-          setValue(
-            (value) =>
-              value.slice(0, cursorPosition - matched.length) +
-              value.slice(cursorPosition)
-          )
+          setValue(editValue)
         }
       }
     },
-    [handleSubmit, setValue, humanToData]
+    [handleSubmit, setValue, handleBackspace]
   )
 
   // Set focus and clear on new chat selected
@@ -235,16 +152,14 @@ export const ChatComposer = (props: ChatComposerProps) => {
     }
     if (chatId !== chatIdRef.current) {
       setValue('')
-      setTrackId(null)
-      setCollectionId(null)
+      reset()
       chatIdRef.current = chatId
     }
-  }, [ref, chatId, chatIdRef, setTrackId, setValue])
+  }, [ref, chatId, chatIdRef, reset, setValue])
 
   const renderChatDisplay = (value: string) => {
-    const regexString = Object.keys(humanToData).join('|')
-    const regex = regexString ? new RegExp(regexString, 'gi') : null
-    if (!regex) {
+    const matches = getMatches(value)
+    if (!matches) {
       const text = splitOnNewline(value)
       return text.map((t, i) => (
         <ComposerText key={i} color='default'>
@@ -252,7 +167,6 @@ export const ChatComposer = (props: ChatComposerProps) => {
         </ComposerText>
       ))
     }
-    const matches = value.matchAll(regex)
     const parts = []
     let lastIndex = 0
     for (const match of matches) {
